requirements.txt
discord.py
asyncio
os